usize is faster

+++ src/day1.rs
@@ -16,10 +16,10 @@ pub fn solve_part1(input: &str) -> i32 {
     return max;
 }
 
-#[aoc(day1, part2)]
-pub fn solve_part2(input: &str) -> i32 {
+#[aoc(day1, part2, adam)]
+pub fn solve_part2(input: &str) -> usize {
     let mut max = vec![0; 3];
-    let mut current_sum = 0;
+    let mut current_sum = 0usize;
     for l in input.lines() {
         if l == "" {
             for i in max.iter_mut() {
@@ -29,7 +29,7 @@ pub fn solve_part2(input: &str) -> i32 {
             }
             current_sum = 0;
         } else {
-            let c: i32 = l.parse().unwrap();
+            let c: usize = l.parse().unwrap();
             current_sum += c;
         }
     }
